Move Connection, Edge etc to graphql scalars
